+++ pkg/cmd/pulumi/history.go
@@ -28,6 +28,7 @@ func newHistoryCmd() *cobra.Command {
 	var stack string
 	var jsonOut bool
 	var showSecrets bool
+	var limit int
 	var cmd = &cobra.Command{
 		Use:        "history",
 		Aliases:    []string{"hist"},
@@ -47,8 +48,9 @@ func newHistoryCmd() *cobra.Command {
 			if err != nil {
 				return err
 			}
+
 			b := s.Backend()
-			updates, err := b.GetHistory(commandContext(), s.Ref())
+			updates, err := b.GetHistory(commandContext(), s.Ref(), limit)
 			if err != nil {
 				return errors.Wrap(err, "getting history")
 			}
@@ -76,5 +78,7 @@ func newHistoryCmd() *cobra.Command {
 		"Show secret values when listing config instead of displaying blinded values")
 	cmd.PersistentFlags().BoolVarP(
 		&jsonOut, "json", "j", false, "Emit output as JSON")
+	cmd.PersistentFlags().IntVarP(
+		&limit, "limit", "l", 0, "Limit the number of entries returned, defaults to all")
 	return cmd
 }