+++ pkg/cmd/pulumi/stack_history.go
@@ -24,6 +24,7 @@ func newStackHistoryCmd() *cobra.Command {
 	var stack string
 	var jsonOut bool
 	var showSecrets bool
+	var limit int
 
 	cmd := &cobra.Command{
 		Use:        "history",
@@ -42,7 +43,7 @@ This command displays data about previous updates for a stack.`,
 				return err
 			}
 			b := s.Backend()
-			updates, err := b.GetHistory(commandContext(), s.Ref())
+			updates, err := b.GetHistory(commandContext(), s.Ref(), limit)
 			if err != nil {
 				return errors.Wrap(err, "getting history")
 			}
@@ -71,6 +72,8 @@ This command displays data about previous updates for a stack.`,
 		"Show secret values when listing config instead of displaying blinded values")
 	cmd.PersistentFlags().BoolVarP(
 		&jsonOut, "json", "j", false, "Emit output as JSON")
+	cmd.PersistentFlags().IntVarP(
+		&limit, "limit", "l", 0, "Limit the number of entries returned, defaults to all")
 	return cmd
 }