+++ sdk/nodejs/x/automation/stack.ts
@@ -415,8 +415,13 @@ export class Stack {
      * Returns a list summarizing all previous and current results from Stack lifecycle operations
      * (up/preview/refresh/destroy).
      */
-    async history(): Promise<UpdateSummary[]> {
-        const result = await this.runPulumiCmd(["history", "--json", "--show-secrets"]);
+    async history(limit?: number): Promise<UpdateSummary[]> {
+        const args = ["history", "--json", "--show-secrets"];
+        if (limit) {
+            args.push("--limit", Math.floor(limit).toString())
+        }
+        const result = await this.runPulumiCmd(args);
+      
         return JSON.parse(result.stdout, (key, value) => {
             if (key === "startTime" || key === "endTime") {
                 return new Date(value);
@@ -425,7 +430,7 @@ export class Stack {
         });
     }
     async info(): Promise<UpdateSummary | undefined> {
-        const history = await this.history();
+        const history = await this.history(1 /*limit*/);
         if (!history || history.length === 0) {
             return undefined;
         }