+++ sdk/python/lib/pulumi/x/automation/stack.py
@@ -464,14 +464,20 @@ class Stack:
             outputs[key] = OutputValue(value=plaintext_outputs[key], secret=secret)
         return outputs
 
-    def history(self) -> List[UpdateSummary]:
+    def history(self,
+        limit: Optional[int] = None) -> List[UpdateSummary]:
         """
         Returns a list summarizing all previous and current results from Stack lifecycle operations
         (up/preview/refresh/destroy).
 
+        :param limit: Limit the number of history entires to retrieve, defaults to all.
+
         :returns: List[UpdateSummary]
         """
-        result = self._run_pulumi_cmd_sync(["history", "--json", "--show-secrets"])
+        args = ["history", "--json", "--show-secrets"]
+        if limit is not None:
+            args.extend(["--limit", str(limit)])
+        result = self._run_pulumi_cmd_sync(args)
         summary_list = json.loads(result.stdout)
 
         summaries: List[UpdateSummary] = []
@@ -495,7 +501,7 @@ class Stack:
 
         :returns: Optional[UpdateSummary]
         """
-        history = self.history()
+        history = self.history(limit=1)
         if not len(history):
             return None
         return history[0]